+++ Code/Assets/Scripts/Skill/Skill_Attack.cs
@@ -80,7 +80,7 @@ namespace Game
                     {
                         if (effect.Config.Priority >= 0)
                         {
-                            long total = dr.Damage * effect.Percent / 100;
+                            double total = dr.Damage * effect.Percent / 100;
                             //Debug.Log("restor:" + total);
                             DoEffect(enemy, this.SelfPlayer, total,0, effect);
                         }
